requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
openpyxl
requests
openai
pytest
pydantic-settings
pdf2image
pytesseract
Pillow
python-dotenv